refactor(app): make the shell own its layout
